First attempt at 64 bit correctness.

+++ coal.c
@@ -99,6 +99,6 @@ void coal_throw (var throwable) {
   } else {
     longjmp(*(jmp_buf *)
 	    utility_stack_pop(&exceptions_s__),
-	    (int) throwable);
+	    (intptr_t) throwable);
   }
 }
